backend/pyproject.toml
[project]
name = "0xrip-backend"
version = "0.1.5"
description = "0xRIP Digital Graveyard Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "requests>=2.32",
    "python-dotenv>=1.1",
    "pydantic>=2.10",
    "sqlalchemy>=2.0",
    "chromadb>=1.0",
    "litellm>=1.55",
    "dspy-ai>=2.6",
]

[project.scripts]
serve = "backend.__main__:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
markers = [
    "integration: tests that call real APIs (requires API keys, skipped by default)",
]
